separate processing and view
